requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.8.4
dill==0.4.1
fastapi==0.136.1
filelock==3.25.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
git-filter-repo==2.47.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.12.0
idna==3.13
Jinja2==3.1.6
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.6.1
numpy==2.4.3
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.2
pandas==3.0.2
pillow==12.1.1
polars==1.40.1
polars-runtime-32==1.40.1
propcache==0.4.1
psutil==7.2.2
pyarrow==24.0.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.27
PyYAML==6.0.3
requests==2.33.1
rich==15.0.0
scipy==1.17.1
setuptools==70.2.0
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
starlette==1.0.0
stringzilla==4.6.0
sympy==1.14.0
tqdm==4.67.3
typer==0.24.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.42
ultralytics-thop==2.0.19
urllib3==2.6.3
uvicorn==0.46.0
xxhash==3.7.0
yarl==1.23.0
